+++ ppr-ui/src/components/mhrRegistration/HomeOwners/index.ts
@@ -1,3 +1,3 @@
-export { default as AddHomeOwnerPerson } from './AddHomeOwnerPerson.vue'
+export { default as AddEditHomeOwnerPerson } from './AddEditHomeOwnerPerson.vue'
 export { default as AddHomeOwnerOrganization } from './AddHomeOwnerOrganization.vue'
 export { default as HomeOwnersTable } from './HomeOwnersTable.vue'
