+++ ppr-ui/src/schemas/index.ts
@@ -1,2 +1,1 @@
 export * from './party-address'
-export * from './organization-address'
